+++ com.timbojimbo.sequencer/Runtime/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("TimboJimboEditor.Sequencer")]
+[assembly: InternalsVisibleTo("TimboJimboTests.Sequencer")]